+++ contracts/vaults/BalanceVaultManager.sol
@@ -15,12 +15,14 @@ import "./BalanceVaultShare.sol";
 import "../utils/ArrayUtils.sol";
 
 interface IBalancePassManager {
-    function getDiscountFromFee(address _user, uint _fee) external view returns (uint, uint);
+    function getDiscountFromFee(address _user, uint256 _fee)
+        external
+        view
+        returns (uint256, uint256);
 }
 
 /// @notice Creates new balance vaults
 contract BalanceVaultManager is Ownable, AccessControl, ReentrancyGuard {
-
     using SafeERC20 for IERC20;
 
     bytes32 public constant MANAGER_ROLE = keccak256("MANAGER_ROLE");
@@ -32,11 +34,11 @@ contract BalanceVaultManager is Ownable, AccessControl, ReentrancyGuard {
     address public nftTemplate;
 
     /// fee on borrowers total amount raised, 2 decimals percent, 100% is 10000
-    uint public feeBorrower;
+    uint256 public feeBorrower;
     /// fee on lenders return in case usdb is used, 2 decimals percent, 100% is 10000
-    uint public feeLenderUsdb;
+    uint256 public feeLenderUsdb;
     /// fee on lenders return in case other token is used, 2 decimals percent, 100% is 10000
-    uint public feeLenderOther;
+    uint256 public feeLenderOther;
 
     /// vetted tokens
     address[] public allowedTokens;
@@ -51,28 +53,26 @@ contract BalanceVaultManager is Ownable, AccessControl, ReentrancyGuard {
 
     struct BalanceVaultDto {
         address vaultAddress;
-        uint index;
+        uint256 index;
         address nftAddress;
-
         string[] ownerInfos;
         string[] ownerContacts;
         address ownerWallet;
-        uint fundingAmount;
-        uint fundraised;
+        uint256 fundingAmount;
+        uint256 fundraised;
         address[] allowedTokens;
-        uint freezeTimestamp;
-        uint repaymentTimestamp;
-        uint apr;
+        uint256 freezeTimestamp;
+        uint256 repaymentTimestamp;
+        uint256 apr;
         bool shouldBeFrozen;
     }
 
     struct BalanceVaultPositionDto {
         address vaultAddress;
-        uint index;
+        uint256 index;
         address nftAddress;
-
         address user;
-        uint[] amounts;
+        uint256[] amounts;
         address[] tokens;
     }
 
@@ -81,7 +81,13 @@ contract BalanceVaultManager is Ownable, AccessControl, ReentrancyGuard {
     /// @param _feeBorrower fee on borrowers total amount raised, 2 decimals percent, 100% is 10000
     /// @param _feeLenderUsdb fee on lenders return in case usdb is used, 2 decimals percent, 100% is 10000
     /// @param _feeLenderOther fee on lenders return in case other token is used, 2 decimals percent, 100% is 10000
-    constructor(address _DAO, address _USDB, uint _feeBorrower, uint _feeLenderUsdb, uint _feeLenderOther) {
+    constructor(
+        address _DAO,
+        address _USDB,
+        uint256 _feeBorrower,
+        uint256 _feeLenderUsdb,
+        uint256 _feeLenderOther
+    ) {
         require(_DAO != address(0));
         DAO = _DAO;
         require(_USDB != address(0));
@@ -132,22 +138,31 @@ contract BalanceVaultManager is Ownable, AccessControl, ReentrancyGuard {
         string[] calldata _ownerInfos,
         string[] calldata _ownerContacts,
         address _ownerWallet,
-        uint _fundingAmount,
+        uint256 _fundingAmount,
         address[] calldata _allowedTokens,
-        uint _freezeTimestamp,
-        uint _repaymentTimestamp,
-        uint _apr
+        uint256 _freezeTimestamp,
+        uint256 _repaymentTimestamp,
+        uint256 _apr
     ) external nonReentrant returns (address _vaultAddress) {
         require(vaultTemplate != address(0), "MISSING_VAULT_TEMPLATE");
         require(nftTemplate != address(0), "MISSING_NFT_TEMPLATE");
 
-        require(_freezeTimestamp < _repaymentTimestamp, "VAULT_FREEZE_SHOULD_BE_BEFORE_PAYOUT");
-        require(_freezeTimestamp > block.timestamp, "VAULT_FREEZE_SHOULD_BE_IN_FUTURE");
+        require(
+            _freezeTimestamp < _repaymentTimestamp,
+            "VAULT_FREEZE_SHOULD_BE_BEFORE_PAYOUT"
+        );
+        require(
+            _freezeTimestamp > block.timestamp,
+            "VAULT_FREEZE_SHOULD_BE_IN_FUTURE"
+        );
 
         require(_ownerInfos.length == 2, "INFOS_MISSING");
 
-        for (uint i = 0; i < _allowedTokens.length; i++) {
-            require(allowedTokensMapping[_allowedTokens[i]], "TOKEN_NOT_ALLOWED");
+        for (uint256 i = 0; i < _allowedTokens.length; i++) {
+            require(
+                allowedTokensMapping[_allowedTokens[i]],
+                "TOKEN_NOT_ALLOWED"
+            );
         }
 
         // EIP1167 clone factory
@@ -155,18 +170,18 @@ contract BalanceVaultManager is Ownable, AccessControl, ReentrancyGuard {
         address nftAddress = Clones.clone(nftTemplate);
 
         VaultParams memory param = VaultParams({
-        ownerInfos : _ownerInfos,
-        ownerContacts : _ownerContacts,
-        ownerWallet : _ownerWallet,
-        nftAddress : nftAddress,
-        fundingAmount : _fundingAmount,
-        allowedTokens : _allowedTokens,
-        freezeTimestamp : _freezeTimestamp,
-        repaymentTimestamp : _repaymentTimestamp,
-        apr : _apr,
-        feeBorrower : feeBorrower,
-        feeLenderUsdb : feeLenderUsdb,
-        feeLenderOther : feeLenderOther
+            ownerInfos: _ownerInfos,
+            ownerContacts: _ownerContacts,
+            ownerWallet: _ownerWallet,
+            nftAddress: nftAddress,
+            fundingAmount: _fundingAmount,
+            allowedTokens: _allowedTokens,
+            freezeTimestamp: _freezeTimestamp,
+            repaymentTimestamp: _repaymentTimestamp,
+            apr: _apr,
+            feeBorrower: feeBorrower,
+            feeLenderUsdb: feeLenderUsdb,
+            feeLenderOther: feeLenderOther
         });
         BalanceVault vault = BalanceVault(_vaultAddress);
         vault.initialize(param);
@@ -181,17 +196,30 @@ contract BalanceVaultManager is Ownable, AccessControl, ReentrancyGuard {
         generatedVaults.push(_vaultAddress);
 
         // remember in history
-        emit VaultCreated(msg.sender, _vaultAddress, vaultTemplate, nftTemplate);
+        emit VaultCreated(
+            msg.sender,
+            _vaultAddress,
+            vaultTemplate,
+            nftTemplate
+        );
     }
 
     /// @notice get amount and fee part from fee
     /// @param _user given user
     /// @param _fee fee to split
     /// @return amount and fee part from given fee
-    function getDiscountFromFee(address _user, uint _fee) external returns (uint, uint) {
+    function getDiscountFromFee(address _user, uint256 _fee)
+        external
+        returns (uint256, uint256)
+    {
         if (balancePassManager == address(0)) return (0, _fee);
 
-        try IBalancePassManager(balancePassManager).getDiscountFromFee(_user, _fee) returns (uint _amount, uint _finalFee) {
+        try
+            IBalancePassManager(balancePassManager).getDiscountFromFee(
+                _user,
+                _fee
+            )
+        returns (uint256 _amount, uint256 _finalFee) {
             return (_amount, _finalFee);
         } catch (bytes memory reason) {
             emit LogBytes(reason);
@@ -205,7 +233,7 @@ contract BalanceVaultManager is Ownable, AccessControl, ReentrancyGuard {
 
     /// @notice get generated vaults length for paging
     /// @return generated vaults length for paging
-    function getGeneratedVaultsLength() external view returns (uint) {
+    function getGeneratedVaultsLength() external view returns (uint256) {
         return generatedVaults.length;
     }
 
@@ -213,28 +241,31 @@ contract BalanceVaultManager is Ownable, AccessControl, ReentrancyGuard {
     /// @param _skip how many items from beginning to skip
     /// @param _limit how many items to return in result which are not blacklisted
     /// @return page of BalanceVaultDto
-    function getGeneratedVaultsPage(uint _skip, uint _limit) external view returns (BalanceVaultDto[] memory) {
+    function getGeneratedVaultsPage(uint256 _skip, uint256 _limit)
+        external
+        view
+        returns (BalanceVaultDto[] memory)
+    {
         if (_skip >= generatedVaults.length) return new BalanceVaultDto[](0);
 
-        uint limit = Math.min(_skip + _limit, generatedVaults.length);
+        uint256 limit = Math.min(_skip + _limit, generatedVaults.length);
         BalanceVaultDto[] memory page = new BalanceVaultDto[](limit);
-        uint index = 0;
-        for (uint i = _skip; i < limit; i++) {
+        uint256 index = 0;
+        for (uint256 i = _skip; i < limit; i++) {
             BalanceVault vault = BalanceVault(generatedVaults[i]);
             // do not send not vetted vaults to the frontend
-            if (!generatedVaultsWhitelist[address (vault)]) continue;
+            if (!generatedVaultsWhitelist[address(vault)]) continue;
 
             string[] memory ownerInfos = new string[](2);
             ownerInfos[0] = vault.ownerName();
             ownerInfos[1] = vault.ownerDescription();
 
             page[index++] = BalanceVaultDto({
-                vaultAddress: address (vault),
+                vaultAddress: address(vault),
                 index: i,
                 nftAddress: address(vault.nft()),
-
-                ownerInfos : ownerInfos,
-                ownerContacts : vault.getOwnerContacts(),
+                ownerInfos: ownerInfos,
+                ownerContacts: vault.getOwnerContacts(),
                 ownerWallet: vault.ownerWallet(),
                 fundingAmount: vault.fundingAmount(),
                 fundraised: vault.fundraised(),
@@ -253,26 +284,33 @@ contract BalanceVaultManager is Ownable, AccessControl, ReentrancyGuard {
     /// @param _skip how many items from beginning to skip
     /// @param _limit how many items to return in result
     /// @return page of BalanceVaultPositionDto
-    function getPositionsPage(address _user, uint _skip, uint _limit) external view returns (BalanceVaultPositionDto[] memory) {
-        if (_skip >= generatedVaults.length) return new BalanceVaultPositionDto[](0);
+    function getPositionsPage(
+        address _user,
+        uint256 _skip,
+        uint256 _limit
+    ) external view returns (BalanceVaultPositionDto[] memory) {
+        if (_skip >= generatedVaults.length)
+            return new BalanceVaultPositionDto[](0);
 
-        uint limit = Math.min(_skip + _limit, generatedVaults.length);
-        BalanceVaultPositionDto[] memory page = new BalanceVaultPositionDto[](limit);
-        uint index = 0;
-        for (uint i = _skip; i < limit; i++) {
+        uint256 limit = Math.min(_skip + _limit, generatedVaults.length);
+        BalanceVaultPositionDto[] memory page = new BalanceVaultPositionDto[](
+            limit
+        );
+        uint256 index = 0;
+        for (uint256 i = _skip; i < limit; i++) {
             BalanceVault vault = BalanceVault(generatedVaults[i]);
             // do not send not vetted vaults to the frontend
-            if (!generatedVaultsWhitelist[address (vault)]) continue;
+            if (!generatedVaultsWhitelist[address(vault)]) continue;
 
-            (uint[] memory _amounts, address[] memory _tokens) = vault.balanceOf(_user);
+            (uint256[] memory _amounts, address[] memory _tokens) = vault
+                .balanceOf(_user);
             // do not send empty positions to the frontend
             if (_amounts.length == 0) continue;
 
             page[index++] = BalanceVaultPositionDto({
-                vaultAddress: address (vault),
+                vaultAddress: address(vault),
                 index: i,
                 nftAddress: address(vault.nft()),
-
                 user: _user,
                 amounts: _amounts,
                 tokens: _tokens
@@ -284,14 +322,22 @@ contract BalanceVaultManager is Ownable, AccessControl, ReentrancyGuard {
     /// @notice when created vault is vetted, operator will add it into the currated list
     /// @param _contractAddress vault CA
     /// @param _add true if addition
-    function modifyGeneratedVaultWhitelist(address _contractAddress, bool _add) external {
+    function modifyGeneratedVaultWhitelist(address _contractAddress, bool _add)
+        external
+    {
         require(hasRole(MANAGER_ROLE, _msgSender()), "MANAGER_ROLE_MISSING");
 
         if (_add) {
-            require(!generatedVaultsWhitelist[_contractAddress], "ALREADY_IN_WHITELIST");
+            require(
+                !generatedVaultsWhitelist[_contractAddress],
+                "ALREADY_IN_WHITELIST"
+            );
             generatedVaultsWhitelist[_contractAddress] = true;
         } else {
-            require(generatedVaultsWhitelist[_contractAddress], "NOT_IN_WHITELIST");
+            require(
+                generatedVaultsWhitelist[_contractAddress],
+                "NOT_IN_WHITELIST"
+            );
             delete generatedVaultsWhitelist[_contractAddress];
         }
     }
@@ -316,21 +362,21 @@ contract BalanceVaultManager is Ownable, AccessControl, ReentrancyGuard {
 
     /// @notice sets fee for total amount raise
     /// @param _feeBorrower fee on borrowers total amount raised, 2 decimals percent, 100% is 10000
-    function setFeeBorrower(uint _feeBorrower) external onlyOwner {
+    function setFeeBorrower(uint256 _feeBorrower) external onlyOwner {
         require(_feeBorrower < 2000, "FEE_TOO_HIGH");
         feeBorrower = _feeBorrower;
     }
 
     /// @notice sets fee for usdb token
     /// @param _feeLenderUsdb fee on lenders return in case usdb is used, 2 decimals percent, 100% is 10000
-    function setFeeLenderUsdb(uint _feeLenderUsdb) external onlyOwner {
+    function setFeeLenderUsdb(uint256 _feeLenderUsdb) external onlyOwner {
         require(_feeLenderUsdb < 3000, "FEE_TOO_HIGH");
         feeLenderUsdb = _feeLenderUsdb;
     }
 
     /// @notice sets fee for other tokens
     /// @param _feeLenderOther fee on lenders return in case other token is used, 2 decimals percent, 100% is 10000
-    function setFeeLenderOther(uint _feeLenderOther) external onlyOwner {
+    function setFeeLenderOther(uint256 _feeLenderOther) external onlyOwner {
         require(_feeLenderOther < 3000, "FEE_TOO_HIGH");
         feeLenderOther = _feeLenderOther;
     }
@@ -339,7 +385,7 @@ contract BalanceVaultManager is Ownable, AccessControl, ReentrancyGuard {
     /// @param _token token CA
     function setAllowedToken(address _token) public onlyOwner {
         address[] memory tokens = new address[](allowedTokens.length + 1);
-        for (uint i = 0; i < allowedTokens.length; i++) {
+        for (uint256 i = 0; i < allowedTokens.length; i++) {
             tokens[i] = allowedTokens[i];
             require(allowedTokens[i] != _token, "TOKEN_ALREADY_USED");
         }
@@ -351,11 +397,15 @@ contract BalanceVaultManager is Ownable, AccessControl, ReentrancyGuard {
     /// @notice remove allowed token
     /// @param _token token to remove with its mapping
     function removeAllowedToken(address _token) external onlyOwner {
-        uint index = ArrayUtils.arrayIndex(allowedTokens, _token, allowedTokens.length);
-        require(index != type(uint).max, "TOKEN_NOT_FOUND");
+        uint256 index = ArrayUtils.arrayIndex(
+            allowedTokens,
+            _token,
+            allowedTokens.length
+        );
+        require(index != type(uint256).max, "TOKEN_NOT_FOUND");
 
         address[] memory tokens = new address[](allowedTokens.length - 1);
-        for (uint i = 0; i < allowedTokens.length; i++) {
+        for (uint256 i = 0; i < allowedTokens.length; i++) {
             if (i < index) tokens[i] = allowedTokens[i];
             else if (i == index) continue;
             else {
@@ -369,7 +419,10 @@ contract BalanceVaultManager is Ownable, AccessControl, ReentrancyGuard {
 
     /// @notice set manager for balance passes
     /// @param _balancePassManager mgr
-    function setBalancePassManager(address _balancePassManager) external onlyOwner {
+    function setBalancePassManager(address _balancePassManager)
+        external
+        onlyOwner
+    {
         balancePassManager = _balancePassManager;
     }
 
@@ -392,5 +445,4 @@ contract BalanceVaultManager is Ownable, AccessControl, ReentrancyGuard {
     function revokeRoleManager(address _account) external {
         revokeRole(MANAGER_ROLE, _account);
     }
-
 }
